First attempt of whois feature
At the moment, the system doesn't work very well. I use a socket to
connect to the registrars and get an answer from them. The answer
provided by IANA or Internic aren't readable or useful so maybe my query
isn't the right one because I don't obtain the same as if I do a query
through their respective websites.

Signed-off-by: Matthieu Berjon <matthieu@berjon.net>