Updating the CLI
I updated the CLI in order to create:

- a main command that can display the main help
- a subcommand called `whois` dedicated to the whois tool

The main command accepts a `version` option to show the current version
of the project and the Python version installed on the machine.

The `whois` subcommand make a simple test on the data to see if it's an
IP address. If true, a request is made and return the JSON data.
Otherwise, the data is considered as a domaine name. Nothing is done for
this part yet.

The domain syntax isn't tested but should be in the future.

Signed-off-by: Matthieu Berjon <matthieu@berjon.net>